Debugged. Changed stuff.
Wrote in some lines for debugging. Wrote some commands to check processes as they happen. line 282 seems like the one with an issue of some sort. Still have to figure out exactly why fiddling with the absolute value for the distance in the X direction has such a great impact on the code.