More debugging on pixel loss
Wrote some code that identifies one of the ways in which we are loosing pixels. That is, given one pixel of the velocity plot, we tried to find the closest point on the other side of the major axis that contained a similar Y value, and a close enough value of X, but such conditions were inexistent. This issue is understood now and identified. This pixel loss is most unavoidable. Now I wrote a bit of code that tells you exactly which pixel was lost and where. However, we are often loosing some pixels along the major axis right at the center of the galaxy. I still do not know why this happens. That is, going from the masked circular gvel plot, we are unable to trasfer some of the pixels in the plot that we should be able to invert into the inverted plot.