Modified version. Runs. Doen't flip yet.
I've been fixing bugs. I got the code to run but the equation system that I'm using is not working. I'm creating imaginary lines and trying to use the stardard equation of a line to create a flip. However the inequality for some reason is never satisfied (even approximately). I'm still looking into it.